Use iCal instaed of Google Calendar API; Send mails with SES; Refactor lambda
