+++ infrastructure/dev/locals.tf
@@ -1,0 +1,5 @@
+locals {
+  ses_sandbox_recipient_emails = [
+    "ekkos22@gmail.com"
+  ]
+}
