+++ infrastructure/dev/main.tf
@@ -27,11 +27,13 @@ resource "aws_lambda_function" "calendar" {
       ENVIRONMENT              = "dev"
       GOOGLE_CREDENTIALS_PARAM = "/calendar/dev/google-credentials-json"
       TOKEN_JSON_PARAM         = "/calendar/dev/token-json"
-      API_KEY_PARAM            = "/ops-master/cloudfront/apikey"
-      SECOND_KEY_PARAM         = "/calendar/dev/payu-second-key"
+
+      API_KEY_PARAM    = "/ops-master/cloudfront/dev/apikey"
+      SECOND_KEY_PARAM = "/calendar/dev/payu-second-key"
 
       ICAL_URL_PARAM       = "/calendar/dev/ical-feed-url"
       SES_FROM_EMAIL_PARAM = "/calendar/dev/ses-from-email"
+      DYNAMODB_TABLE_NAME  = aws_dynamodb_table.calendar_events.name
     }
   }
 }
