+++ infrastructure/dev/ses.tf
@@ -1,5 +1,22 @@
-resource "aws_ses_email_identity" "calendar_email" {
+resource "aws_ses_email_identity" "calendar" {
   email = var.ses_from_email
+}
+
+resource "aws_ses_email_identity" "sandbox_recipients" {
+  for_each = toset(local.ses_sandbox_recipient_emails)
+  email    = each.value
+}
+
+resource "aws_ses_domain_identity" "calendar" {
+  domain = "ops-master.com"
+}
+
+resource "aws_route53_record" "opsmaster_amazonses_verification_record" {
+  zone_id = data.aws_route53_zone.opsmaster.zone_id
+  name    = "_amazonses.ops-master.com"
+  type    = "TXT"
+  ttl     = "600"
+  records = [aws_ses_domain_identity.calendar.verification_token]
 }
 
 resource "aws_ses_configuration_set" "calendar" {
