+++ infrastructure/dev/terraform.tfvars
@@ -1,3 +1,3 @@
 project_name   = "calendar"
 environment    = "dev"
-ses_from_email = "noreply@ops-master.com"
+ses_from_email = "opsmaster.trainings@gmail.com"
