+++ infrastructure/google_lambda_layer/requirements.txt
@@ -2,3 +2,4 @@ google-api-python-client==2.39.0
 google-auth==2.6.6
 google-auth-httplib2==0.1.0
 google-auth-oauthlib==0.4.6
+icalendar
