+++ requirements.txt
@@ -1,7 +1,8 @@
 boto3
+icalendar
 google-auth
 google-auth-oauthlib
 google-auth-httplib2
 google-api-python-client
 requests
-hashlib
+# hashlib
